requirements/base.txt
# django web framework
django==1.3.1
# required to get docs in the django admin
docutils==0.8.1

# Requests - simpler http requests, for interacting with nodes
requests==0.10.2

# South - database/model migrations
South==0.7.3


# users
django-registration
